+++ tests/android/app/src/main/java/scratch/ts/MainActivity.java
@@ -1,0 +1,5 @@
+package scratch.ts;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}